// shared/src/ProjectTreeTable/types/project.ts
export type FolderType = {
  name: string
  original_name?: string
  shortName?: string
  icon?: string
}
export type TaskType = {
  name: string
  original_name?: string
  shortName?: string
  color?: string
  icon?: string
}